Numpy fix

+++ trm/pgplot/_pgplot.pyx
@@ -5,7 +5,7 @@ cimport cpgplot
 
 FTYPE = np.float32
 DTYPE = np.float64
-ITYPE = np.int
+ITYPE = np.int32
 ctypedef np.float32_t FTYPE_t
 ctypedef np.float64_t DTYPE_t
 ctypedef np.int_t ITYPE_t
